requirements.txt
# ============================================================
# Requirements — Generative AI using Python 2026
# Install: uv add -r requirements.txt
#          OR: pip install -r requirements.txt
# ============================================================

# Core AI SDKs
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Environment & utils
python-dotenv>=1.0.0
httpx>=0.27.0
pypdf>=5.0.0

# Tokenisation & embeddings 
tiktoken>=0.7.0
sentence-transformers>=3.0.0

# Thai NLP 
pythainlp>=5.0.0

# Vector database 
qdrant-client>=1.10.0
faiss-cpu>=1.8.0

# BM25 hybrid search 
rank-bm25>=0.2.2

# RAG evaluation 
ragas>=0.1.0
datasets>=2.20.0

# Structured output 
instructor>=1.3.0
pydantic>=2.7.0

# Deep learning 
torch>=2.3.0
numpy>=1.26.0
matplotlib>=3.8.0

# Fine-tuning & alignment 
transformers>=4.44.0
peft>=0.12.0
trl>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.33.0

# Agents & frameworks 
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# API server 
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0

# LLMOps & observability
langfuse>=2.50.0

# NLI hallucination detection 
# Note: transformers already listed above
